--- src/main/java/com/grepp/teamnotfound/infra/error/exception/GlobalExceptionHandler.java
@@ -1,22 +1,0 @@
-package com.grepp.teamnotfound.infra.error.exception;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import java.time.LocalDateTime;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(CommonException.class)
-    public ResponseEntity<ErrorResponse> handleCommonException(CommonException e) {
-
-        return ResponseEntity
-                .status(e.getErrorCode().getStatus())
-                .body(new ErrorResponse(
-                        e.getErrorCode().getStatus(),
-                        e.getErrorCode().getCode(),
-                        e.getErrorCode().getMessage(),
-                        LocalDateTime.now()));
-    }
-}
