+++ src/main/java/com/grepp/teamnotfound/infra/error/exception/advice/GlobalExceptionHandler.java
@@ -1,0 +1,62 @@
+package com.grepp.teamnotfound.infra.error.exception.advice;
+
+import com.grepp.teamnotfound.infra.error.exception.CommonException;
+import com.grepp.teamnotfound.infra.error.exception.dto.ErrorResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import java.time.LocalDateTime;
+
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(CommonException.class)
+    public ResponseEntity<ErrorResponse> handleCommonException(CommonException e) {
+
+        return ResponseEntity
+                .status(e.getErrorCode().getStatus())
+                .body(new ErrorResponse(
+                        e.getErrorCode().getStatus(),
+                        e.getErrorCode().getCode(),
+                        e.getErrorCode().getMessage(),
+                        LocalDateTime.now()));
+    }
+
+    // 유효성 검증 @Valid Exception catch
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<ErrorResponse> handleValidationExceptions(MethodArgumentNotValidException e) {
+
+        BindingResult bindingResult = e.getBindingResult();
+
+        String errorMessage = bindingResult.getFieldErrors().getFirst().getDefaultMessage();
+
+        return ResponseEntity
+                .status(HttpStatus.BAD_REQUEST)
+                .body(new ErrorResponse(
+                        HttpStatus.BAD_REQUEST.value(),
+                        "VALIDATION_ERROR",
+                        errorMessage,
+                        LocalDateTime.now()));
+    }
+
+    // 그 외 모든 Exception catch
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleException(Exception e) {
+        log.error("Exception occurs: {}", e.getMessage(), e);
+
+        ErrorResponse response = new ErrorResponse(
+                HttpStatus.INTERNAL_SERVER_ERROR.value(),
+                "INTERNAL_SERVER_ERROR",
+                "내부 서버에 예상치 못한 오류가 발생했습니다.",
+                LocalDateTime.now());
+
+        return ResponseEntity
+                .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                .body(response);
+    }
+}
